+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Dizznem Bot Changelog
 
 ## [2.1.2] - 2026-XX-XX
+## Added
+- New admin commands.
+    - $addmoney (adds money to a user).
+    - $resetcooldown (resets cooldowns, e.g. $daily and $weekly).
+
 ### Changed
 - Formatting of AI cooldown / cap messages.
 - Level presentation for both $level and $profile.